Intermed step - create Step3-To_ISMIP_Grid
Pushing this before I mess around with it more. Not functional. This notebook applies Donald's extrapolation/propagation scheme to get offshore ocean TF to each point in the ISMIP grid.  Tested on EN4, having problems with the grid in both pyproj and rioxarray approaches. Now will test with CESM2 and work around that, as it's what we actually need to process at the moment